+++ test/tucan/view_test.exs
@@ -8,4 +8,10 @@ defmodule Tucan.ViewTest do
 
     assert Tucan.View.set_background(Vl.new(), "green") == expected
   end
+
+  test "set_view_background/2" do
+    vl = Tucan.View.set_view_background(Vl.new(), "red")
+
+    assert get_in(vl.spec, ["view"]) == %{"fill" => "red"}
+  end
 end